+++ .gitignore
@@ -24,3 +24,4 @@ prof.out
 *.out
 *.pt
 /Scripts/
+/hdf5_scripts/
